Error handling and env variable added to fetchStations
